pyproject.toml
[build-system]
requires = ["poetry==0.12.10"]
build-backend = "poetry.masonry.api"


[tool.poetry]
name = "fastpages"
description = "Fast page SaaS project"
version = "0.1.0"
readme = "README.md"
authors = []


[tool.poetry.dependencies]
python = "^3.6"
django = "^2.1"
django-split-settings = "^0.3"
django-axes = "^4.5"
django-csp = "^3.4"
django-referrer-policy = "^1.0"
psycopg2 = "^2.7"
gunicorn = "^19.9"
python-decouple = "^3.1"
dump-env = "^0.2"
bcrypt = "^3.1"
django-tenant-schemas = "^1.9"
django-tenant-users = "^0.3.11"
django-rest-auth = "^0.9.3"
django-cors-headers = "^2.4"
djangorestframework_simplejwt = "^3.3"
drf-rw-serializers = "^1.0"
pillow = "^5.4"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^1.11"
django-querycount = "^0.7"
nplusone = "^1.0"
wemake-python-styleguide = "^0.6"
flake8-pytest = "^1.3"
pytest = "^4.0"
pytest-django = "^3.4"
pytest-cov = "^2.6"
pytest-randomly = "^1.2"
pytest-deadfixtures = "^2.1"
pytest-testmon = "^0.9"
pytest-timeout = "^1.3"
pre-commit = "^1.13"
xenon = "^0.5"
mypy = "^0.650"
ipython = "^7.2"
sphinx = "^1.8"
sphinx-autodoc-typehints = "^1.6"
doc8 = "^0.8"
yamllint = "^1.13"
safety = "^1.8"
dotenv-linter = "^0.1"
polint = "^0.4"
